tree-sitter: rename tree_sitter => treesitter for consistency
