--- src/nvim/lua/tree_sitter.h
@@ -1,10 +1,0 @@
-#ifndef NVIM_LUA_TREE_SITTER_H
-#define NVIM_LUA_TREE_SITTER_H
-
-#include "tree_sitter/api.h"
-
-#ifdef INCLUDE_GENERATED_DECLARATIONS
-# include "lua/tree_sitter.h.generated.h"
-#endif
-
-#endif  // NVIM_LUA_TREE_SITTER_H
